pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.2"
pydantic = "^2.8.2"
python-telegram-bot = {version = "^21.0.1", extras = ["job-queue"]}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
pyyaml = "^6.0.1"
apscheduler = "^3.10.4"
jinja2 = "^3.1.3"
minio = "^7.2.5"
SQLAlchemy = "^2.0.28"
asyncpg = "^0.29.0"
filetype = "^1.2.0"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
oic = "^1.6.1"
pillow = "^10.3.0"
pyzbar = "^0.1.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
pre-commit = "^3.8.0"
pyright = "^1.1.378"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"



[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "E501",  # Line too long
  "D",  # pydocstyle
  "S",  # flake8-bandit
  "W",  # pycodestyle warnings (покрываются форматированием)
  "EM", "TRY003",  # Передача строк в конструктор исключений
  "TD", "FIX",  # flake8-todos, flake8-fixme
  "TCH",  # flake8-type-checking
  "ANN101", "ANN102",  # Missing type annotation for `self/cls`
  "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
  "RUF001", "RUF002", "RUF003",  # ambiguous character warnings (Не работают с кириллицей)
  "COM812", "ISC001",  # following rules may cause conflicts when used with the formatter
  "PGH003",  # Use specific rule codes when ignoring type issues
  "BLE001",  # Do not catch blind exception: `Exception`
  "ERA",  # Found commented-out code
  "PLR0913",  # Too many arguments in function definition
  "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.per-file-ignores]
"alembic/**/*.py" = [
  "INP", "E402"
]



[tool.pyright]
exclude = [".venv/", "**/__pycache__", ".*cache*/", "__*/", "alembic/versions/"]
venvPath = "."
venv = ".venv"

reportUnnecessaryTypeIgnoreComment = true
